pyproject.toml
[tool.poetry]
name = "plaid2firefly"
version = "0.1.0"
description = "A connector between Plaid and Firefly III"
authors = ["Erwin Douna <e.douna@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "plaid2firefly" }]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = { extras = ["standard"], version = "^0.29.0" }
httpx = "^0.27.0"
python-dotenv = "^1.0.1"
pydantic = "^2.6.0"
plaid-python = "^16.0.0"  # Uncomment if using Plaid SDK
yarl = "^1.20.0"
requests = "^2.32.3"
aiohttp = "^3.11.18"
asyncio = "^3.4.3"


[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
isort = "^5.13.2"
pytest = "^8.1.1"
pytest-asyncio = "^0.23.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
